+++ test/spanish_doc_test.exs
@@ -16,4 +16,17 @@ defmodule SpanishDocTest do
     {:ok, _type, nie} = SpanishDoc.parse("y7136855s")
     assert to_string(nie) == "Y7136855S"
   end
+
+  test "Ecto.Changeset usings SpanishDoc.EctoType returns an error for invalid data" do
+    types = %{field: SpanishDoc.EctoType}
+    changeset = Ecto.Changeset.cast({%{}, types}, %{field: "world"}, [:field])
+    refute changeset.valid?
+  end
+
+  test "Ecto.Changeset usings SpanishDoc.EctoType returns a parsed field for valid data" do
+    types = %{field: SpanishDoc.EctoType}
+    changeset = Ecto.Changeset.cast({%{}, types}, %{field: "y7136855s"}, [:field])
+    assert {:ok, "Y7136855S"} == Ecto.Changeset.fetch_change(changeset, :field)
+    assert changeset.valid?
+  end
 end